# docs/swagger.yaml
definitions:
  controller.ByNum:
    properties:
      data:
        $ref: '#/definitions/entity.Candidat'
    type: object
  controller.CandidatInput:
    properties:
      gender:
        type: string
      name:
        type: string
      num:
        type: integer
    type: object
  controller.Createresponse:
    properties:
      avatar:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      num:
        type: integer
      vote_number:
        type: integer
    type: object
  controller.DataLists:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.Candidat'
        type: array
    type: object
  entity.Candidat:
    properties:
      avatar:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      num:
        type: integer
      vote_number:
        type: integer
    type: object
info:
  contact: {}
paths:
  /candidat/men:
    get:
      consumes:
      - application/json
      description: find all men candidat
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DataLists'
      summary: find all men
      tags:
      - candidat
  /candidat/woman:
    get:
      consumes:
      - application/json
      description: find some specific candidat
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ByNum'
      summary: find specific candidat
      tags:
      - candidat
    post:
      consumes:
      - application/json
      description: find some specific candidat
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.CandidatInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Createresponse'
      summary: find specific candidat
      tags:
      - candidat
swagger: "2.0"